task_manager/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==5.0.7
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-elasticsearch-dsl==7.4
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.7
django-timezone-field==7.0
djangorestframework==3.15.2
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
flower==2.0.1
humanize==4.10.0
kombu==5.3.7
packaging==24.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
vine==5.1.0
wcwidth==0.2.13